feat (AbstractJwtTokenProvider) : separate redundant code
